Refactor: User interface 구분
