+++ src/constants/types/User.ts
@@ -1,0 +1,14 @@
+export type UserRole = "ADMIN" | "USER";
+
+export interface User {
+  id: number;
+  githubId: string;
+  profileImage: string;
+  introduction: string;
+  email: string;
+  nickName: string;
+  githubUsername: string;
+  createdAt: string;
+  updatedAt: string;
+  role: UserRole;
+}
