+++ src/contexts/AuthContext.tsx
@@ -1,19 +1,5 @@
 import { createContext, useContext } from "react";
-
-export type UserRole = "ADMIN" | "USER";
-
-interface User {
-  id: number;
-  githubId: string;
-  profileImage: string;
-  introduction: string;
-  email: string;
-  nickName: string;
-  githubUsername: string;
-  createdAt: string;
-  updatedAt: string;
-  role: UserRole;
-}
+import { User } from "../constants/types/User";
 
 interface AuthContextType {
   user: User | null;
